requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.8
django-celery-beat==2.0.0
django-cors-headers==3.4.0
django-timezone-field==4.0
djangorestframework==3.11.0
future==0.18.2
kombu==4.6.11
psycopg2-binary==2.8.5
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
